Move queries and commands to Database module
